R peralva/pythonbasics (#433)
* Translated Java: Basics

Translated to Brazilian Portuguese.

* Deleting Java Basics

* lets review python basics

* Update reading-from-console.md

Changed valor to value

* Update using-trinket.md

Changed executar para run

* Update variables.md

Added `` to word double

* Update conditional-statements.md

Added (statement) in front of the work declaração.

* Update conditional-statements.md

Added (statement) in front of the word declaração.

* Update conditional-statements.md

Replaced Declaração to Statement.

* Update conditional-statements.md

Replaced in the row 48 the word declaração to declarações.

* Update conditional-statements.md

In the row 59 replaced instrução2 to declaração2

* Update conditional-statements.md

In the row 64 replaced hungry to comFome

* Update else-if-statement.md

In the row 13 replaced ElifKeyword to A palavra-chaveElif

* Update else-if-statement.md

From de row 17 to 22, It was replaced declaração(ões) to each declaração, as below:

ifexpressão1:
 declaração1
 elifexpressão2:
 declaração2
 else:
 declaração3

* Update else-if-statement.md

Translated row 49 to Portuguese

* Update else-if-statement.md

Translated the row 77 to Portuguese - Python:
 Fundamentos-booleanos

* Update else-if-statement.md

In the row 79 replaced booleans to booleanos

* Update loops.md

In the row 58 replaced the word contar to count

* Update loops.md

In the row 83 replaced: 'Isso diz que a contagem é
 iguala 10' to 'Isso diz que se count for igual a 10'

* Update loops.md

In the row 87 replaced contar to count

* Update nested-loops.md

In the rows 18 - 21 translated expression and statement

* Update nested-loops.md

Replaced the word instrução to declaração in the row 23

* Update _index.md

In the row 18 replaced high code to high level and low code to low level.

* Update nested-loops.md

In the row 25 replaced instrução to declaração

* Update nested-loops.md

In the row 33 replaced enquanto to while

* Update nested-loops.md

In the row 35:
replaced passa to Passando

In the row 37:
Replaced enquanto to while

* Update nested-loops.md

In the row 39:
Replaced passa to Passando

In the rows 48 - 65:
Replaced passa to Passando

In the rows 86 and 87:
Replaced range to intervalo

In the row 91:
Added (rangeA)and (rangeB) after words intervaloA and intervalo to explain better.

In the row 93:
Replaced instrução to declaração.

---------

Co-authored-by: ThaisCristinaMurca <145729632+ThaisCristinaMurca@users.noreply.github.com>